refactor: move database.py to src/database/ with repository pattern
